+++ stacks-node/src/event_dispatcher.rs
@@ -1,5 +1,5 @@
 // Copyright (C) 2013-2020 Blockstack PBC, a public benefit corporation
-// Copyright (C) 2020-2025 Stacks Open Internet Foundation
+// Copyright (C) 2020-2026 Stacks Open Internet Foundation
 //
 // This program is free software: you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
@@ -16,14 +16,12 @@
 
 use std::collections::hash_map::Entry;
 use std::collections::{HashMap, HashSet};
-use std::fmt;
 use std::path::PathBuf;
 #[cfg(test)]
 use std::sync::mpsc::channel;
-#[cfg(test)]
-use std::sync::LazyLock;
 use std::sync::{Arc, Mutex};
-use std::thread::sleep;
+#[cfg(test)]
+use std::sync::{LazyLock, Weak};
 use std::time::{Duration, SystemTime};
 
 use clarity::vm::costs::ExecutionCost;
@@ -31,7 +29,6 @@ use clarity::vm::events::{FTEventType, NFTEventType, STXEventType};
 use clarity::vm::types::{AssetIdentifier, QualifiedContractIdentifier};
 #[cfg(any(test, feature = "testing"))]
 use lazy_static::lazy_static;
-use rand::Rng;
 use serde_json::json;
 use stacks::burnchains::{PoxConstants, Txid};
 use stacks::chainstate::burn::ConsensusHash;
@@ -54,19 +51,17 @@ use stacks::net::api::postblock_proposal::{
     BlockValidateOk, BlockValidateReject, BlockValidateResponse,
 };
 use stacks::net::atlas::{Attachment, AttachmentInstance};
-use stacks::net::http::HttpRequestContents;
-use stacks::net::httpcore::{send_http_request, StacksHttpRequest};
 use stacks::net::stackerdb::StackerDBEventDispatcher;
 #[cfg(any(test, feature = "testing"))]
 use stacks::util::tests::TestFlag;
 use stacks_common::bitvec::BitVec;
 use stacks_common::types::chainstate::{BlockHeaderHash, BurnchainHeaderHash, StacksBlockId};
-use stacks_common::types::net::PeerHost;
 use url::Url;
 
 mod db;
 mod payloads;
 mod stacker_db;
+mod worker;
 
 use db::EventDispatcherDbConnection;
 use payloads::*;
@@ -77,6 +72,7 @@ pub use payloads::{
 pub use stacker_db::StackerDBChannel;
 
 use crate::event_dispatcher::db::PendingPayload;
+use crate::event_dispatcher::worker::{EventDispatcherResult, EventDispatcherWorker};
 
 #[cfg(test)]
 mod tests;
@@ -87,48 +83,23 @@ lazy_static! {
     pub static ref TEST_SKIP_BLOCK_ANNOUNCEMENT: TestFlag<bool> = TestFlag::default();
 }
 
-#[derive(Debug)]
+#[derive(Debug, thiserror::Error)]
 enum EventDispatcherError {
-    SerializationError(serde_json::Error),
-    HttpError(std::io::Error),
-    DbError(stacks::util_lib::db::Error),
+    #[error("Serialization error: {0}")]
+    SerializationError(#[from] serde_json::Error),
+    #[error("HTTP error: {0}")]
+    HttpError(#[from] std::io::Error),
+    #[error("Database error: {0}")]
+    DbError(#[from] stacks::util_lib::db::Error),
+    #[error("Channel receive error: {0}")]
+    RecvError(#[from] std::sync::mpsc::RecvError),
+    #[error("Channel send error: {0}")]
+    SendError(String), // not capturing the underlying because it's a generic type
 }
 
-impl fmt::Display for EventDispatcherError {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match *self {
-            EventDispatcherError::SerializationError(ref e) => fmt::Display::fmt(e, f),
-            EventDispatcherError::HttpError(ref e) => fmt::Display::fmt(e, f),
-            EventDispatcherError::DbError(ref e) => fmt::Display::fmt(e, f),
-        }
-    }
-}
-
-impl core::error::Error for EventDispatcherError {
-    fn cause(&self) -> Option<&dyn core::error::Error> {
-        match *self {
-            EventDispatcherError::SerializationError(ref e) => Some(e),
-            EventDispatcherError::HttpError(ref e) => Some(e),
-            EventDispatcherError::DbError(ref e) => Some(e),
-        }
-    }
-}
-
-impl From<serde_json::Error> for EventDispatcherError {
-    fn from(value: serde_json::Error) -> Self {
-        EventDispatcherError::SerializationError(value)
-    }
-}
-
-impl From<stacks::util_lib::db::Error> for EventDispatcherError {
-    fn from(value: stacks::util_lib::db::Error) -> Self {
-        EventDispatcherError::DbError(value)
-    }
-}
-
-impl From<std::io::Error> for EventDispatcherError {
-    fn from(value: std::io::Error) -> Self {
-        EventDispatcherError::HttpError(value)
+impl<T> From<std::sync::mpsc::SendError<T>> for EventDispatcherError {
+    fn from(value: std::sync::mpsc::SendError<T>) -> Self {
+        EventDispatcherError::SendError(format!("{value}"))
     }
 }
 
@@ -213,6 +184,10 @@ pub struct EventDispatcher {
     pub stackerdb_channel: Arc<Mutex<StackerDBChannel>>,
     /// Path to the database where pending payloads are stored.
     db_path: PathBuf,
+    /// The worker thread that performs the actuall HTTP requests so that they don't block
+    /// the main operation of the node. It's wrapped in an `Arc` only to make some test helpers
+    /// work (see `ALL_WORKERS`); in release code it wouldn't be necessary.
+    worker: Arc<EventDispatcherWorker>,
 }
 
 /// This struct is used specifically for receiving proposal responses.
@@ -237,7 +212,9 @@ impl ProposalCallbackReceiver for ProposalCallbackHandler {
 
         for observer in self.observers.iter() {
             self.dispatcher
-                .dispatch_to_observer(observer, &response, PATH_PROPOSAL_RESPONSE);
+                .dispatch_to_observer(observer, &response, PATH_PROPOSAL_RESPONSE)
+                .unwrap()
+                .wait_until_complete();
         }
     }
 }
@@ -409,11 +386,63 @@ impl BlockEventDispatcher for EventDispatcher {
     }
 }
 
+/// During integration tests, the `test_observer` needs to ensure that all events
+/// that were triggered have actually been delivered, before it can pass on the
+/// captured data. To make that work, during test we store weak references to
+/// all the workers and make it possible to wait for all of them to catch up
+/// in a single function call (see `catch_up_all_event_dispatchers`).
+#[cfg(test)]
+static ALL_WORKERS: Mutex<Vec<Weak<EventDispatcherWorker>>> = Mutex::new(Vec::new());
+
+#[cfg(test)]
+pub fn catch_up_all_event_dispatchers() {
+    let mut results = Vec::new();
+    let mut guard = ALL_WORKERS.lock().unwrap();
+
+    // remove all items that have been dropped; call .noop() the rest
+    guard.retain_mut(|w| {
+        let Some(worker) = w.upgrade() else {
+            return false;
+        };
+        results.push(worker.noop().unwrap());
+        return true;
+    });
+    // unlock the mutex
+    drop(guard);
+
+    // block until all workers have caught up
+    for result in results {
+        result.wait_until_complete();
+    }
+}
+
 impl EventDispatcher {
+    /// The default behavior is to create a non-blocking dispatcher with a
+    /// queue size of 1,000. Note however that the default *node* configuration
+    /// is to always block (i.e. an effective queue size of 0).
+    ///
+    /// See the `event_dispatcher_blocking` and `event_dispatcher_queue_size`
+    /// config values.
     pub fn new(working_dir: PathBuf) -> EventDispatcher {
+        Self::new_with_custom_queue_size(working_dir, 1_000)
+    }
+    /// The queue size specifies how many events may be in-flight without
+    /// blocking the calling thread when sending additional events. A value
+    /// of 0 means they always block.
+    pub fn new_with_custom_queue_size(working_dir: PathBuf, queue_size: usize) -> EventDispatcher {
         let mut db_path = working_dir;
         db_path.push("event_observers.sqlite");
         EventDispatcherDbConnection::new(&db_path).expect("Failed to initialize database");
+
+        let worker = EventDispatcherWorker::new(db_path.clone(), queue_size)
+            .expect("Failed to start worker thread");
+
+        let worker = Arc::new(worker);
+
+        #[cfg(test)]
+        {
+            ALL_WORKERS.lock().unwrap().push(Arc::downgrade(&worker));
+        }
 
         EventDispatcher {
             stackerdb_channel: Arc::new(Mutex::new(StackerDBChannel::new())),
@@ -430,7 +459,17 @@ impl EventDispatcher {
             stackerdb_observers_lookup: HashSet::new(),
             block_proposal_observers_lookup: HashSet::new(),
             db_path,
+            worker,
         }
+    }
+
+    /// Sends a noop task to the worker and waits until its completion is acknowledged.
+    /// This has the effect that all payloads that have been submitted before this point
+    /// are also done, which is a useful thing to wait for in some tests where you want
+    /// to assert on certain event deliveries.
+    #[cfg(test)]
+    pub fn catch_up(&self) {
+        self.worker.noop().unwrap().wait_until_complete();
     }
 
     pub fn process_burn_block(
@@ -643,7 +682,7 @@ impl EventDispatcher {
                 );
 
                 // Send payload
-                self.dispatch_to_observer(
+                self.dispatch_to_observer_or_log_error(
                     &self.registered_observers[observer_id],
                     &payload,
                     PATH_BLOCK_PROCESSED,
@@ -1046,8 +1085,9 @@ impl EventDispatcher {
         event_observer
     }
 
-    /// Process any pending payloads in the database.
-    /// This is called when the event dispatcher is first instantiated.
+    /// Process any pending payloads in the database. This is meant to be called at startup, in order to
+    /// handle anything that was enqueued but not sent before shutdown. This method blocks until all
+    /// requests are made (or, if the observer is no longer registered, removed from the DB).
     pub fn process_pending_payloads(&self) {
         let conn =
             EventDispatcherDbConnection::new(&self.db_path).expect("Failed to initialize database");
@@ -1068,9 +1108,7 @@ impl EventDispatcher {
         );
 
         for PendingPayload {
-            id,
-            mut request_data,
-            ..
+            id, request_data, ..
         } in pending_payloads
         {
             info!(
@@ -1112,18 +1150,22 @@ impl EventDispatcher {
 
             // If the timeout configuration for this observer is different from what it was
             // originally, the updated config wins.
-            request_data.timeout = observer.timeout;
-
-            self.make_http_request_and_delete_from_db(&request_data, observer.disable_retries, id);
+            self.worker
+                .initiate_send(id, observer.disable_retries, Some(observer.timeout))
+                .expect("failed to dispatch pending event payload to worker thread")
+                .wait_until_complete();
         }
     }
 
+    /// A successful result from this method only indicates that that payload was successfully
+    /// enqueued, not that the HTTP request was actually made. If you need to wait until that's
+    /// the case, call `wait_until_complete()` on the `EventDispatcherResult`.
     fn dispatch_to_observer(
         &self,
         event_observer: &EventObserver,
         payload: &serde_json::Value,
         path: &str,
-    ) {
+    ) -> Result<EventDispatcherResult, EventDispatcherError> {
         let full_url = Self::get_full_url(event_observer, path);
         let bytes = match Self::get_payload_bytes(payload) {
             Ok(bytes) => bytes,
@@ -1131,7 +1173,7 @@ impl EventDispatcher {
                 error!(
                     "Event dispatcher: failed to serialize payload"; "path" => path, "error" => ?err
                 );
-                return;
+                return Err(err);
             }
         };
 
@@ -1143,81 +1185,25 @@ impl EventDispatcher {
 
         let id = self.save_to_db(&data);
 
-        self.make_http_request_and_delete_from_db(&data, event_observer.disable_retries, id);
+        self.worker
+            .initiate_send(id, event_observer.disable_retries, None)
     }
 
-    fn make_http_request(
-        data: &EventRequestData,
-        disable_retries: bool,
-    ) -> Result<(), EventDispatcherError> {
-        debug!(
-            "Event dispatcher: Sending payload"; "url" => &data.url, "bytes" => data.payload_bytes.len()
-        );
-
-        let url = Url::parse(&data.url)
-            .unwrap_or_else(|_| panic!("Event dispatcher: unable to parse {} as a URL", data.url));
-
-        let host = url.host_str().expect("Invalid URL: missing host");
-        let port = url.port_or_known_default().unwrap_or(80);
-        let peerhost: PeerHost = format!("{host}:{port}")
-            .parse()
-            .unwrap_or(PeerHost::DNS(host.to_string(), port));
-
-        let mut backoff = Duration::from_millis(100);
-        let mut attempts: i32 = 0;
-        // Cap the backoff at 3x the timeout
-        let max_backoff = data.timeout.saturating_mul(3);
-
-        loop {
-            let mut request = StacksHttpRequest::new_for_peer(
-                peerhost.clone(),
-                "POST".into(),
-                url.path().into(),
-                HttpRequestContents::new().payload_json_bytes(Arc::clone(&data.payload_bytes)),
-            )
-            .unwrap_or_else(|_| panic!("FATAL: failed to encode infallible data as HTTP request"));
-            request.add_header("Connection".into(), "close".into());
-            match send_http_request(host, port, request, data.timeout) {
-                Ok(response) => {
-                    if response.preamble().status_code == 200 {
-                        debug!(
-                            "Event dispatcher: Successful POST"; "url" => %url
-                        );
-                        break;
-                    } else {
-                        error!(
-                            "Event dispatcher: Failed POST"; "url" => %url, "response" => ?response.preamble()
-                        );
-                    }
-                }
-                Err(err) => {
-                    warn!(
-                        "Event dispatcher: connection or request failed to {host}:{port} - {err:?}";
-                        "backoff" => ?backoff,
-                        "attempts" => attempts
-                    );
-                    if disable_retries {
-                        warn!("Observer is configured in disable_retries mode: skipping retry of payload");
-                        return Err(err.into());
-                    }
-                    #[cfg(test)]
-                    if TEST_EVENT_OBSERVER_SKIP_RETRY.get() {
-                        warn!("Fault injection: skipping retry of payload");
-                        return Err(err.into());
-                    }
-                }
-            }
-
-            sleep(backoff);
-            let jitter: u64 = rand::thread_rng().gen_range(0..100);
-            backoff = std::cmp::min(
-                backoff.saturating_mul(2) + Duration::from_millis(jitter),
-                max_backoff,
-            );
-            attempts = attempts.saturating_add(1);
+    /// This fire-and-forget version of `dispatch_to_observer` logs any error from enqueueing the
+    /// request, and does not give you a way to wait for blocking until it's sent. If you need
+    /// more control, use `dispatch_to_observer()` directly and handle the result yourself.
+    ///
+    /// This method exists because we generally don't want the event dispatcher to interrupt the node's
+    /// processing.
+    fn dispatch_to_observer_or_log_error(
+        &self,
+        event_observer: &EventObserver,
+        payload: &serde_json::Value,
+        path: &str,
+    ) {
+        if let Err(err) = self.dispatch_to_observer(event_observer, payload, path) {
+            error!("Event dispatcher: Failed to enqueue payload for sending to observer: {err:?}");
         }
-
-        Ok(())
     }
 
     fn get_payload_bytes(payload: &serde_json::Value) -> Result<Arc<[u8]>, EventDispatcherError> {
@@ -1245,53 +1231,12 @@ impl EventDispatcher {
         conn.insert_payload_with_retry(data, SystemTime::now())
     }
 
-    fn make_http_request_and_delete_from_db(
-        &self,
-        data: &EventRequestData,
-        disable_retries: bool,
-        id: i64,
-    ) {
-        let http_result = Self::make_http_request(data, disable_retries);
-
-        if let Err(err) = http_result {
-            // log but continue
-            error!("EventDispatcher: dispatching failed"; "url" => data.url.clone(), "error" => ?err);
-        }
-
-        #[cfg(test)]
-        if TEST_EVENT_OBSERVER_SKIP_RETRY.get() {
-            warn!("Fault injection: skipping deletion of payload");
-            return;
-        }
-
-        // We're deleting regardless of result -- if retries are disabled, that means
-        // we're supposed to forget about it in case of failure. If they're not disabled,
-        // then we wouldn't be here in case of failue, because `make_http_request` retries
-        // until it's successful (with the exception of the above fault injection which
-        // simulates a shutdown).
-        let deletion_result = self.delete_from_db(id);
-
-        if let Err(e) = deletion_result {
-            error!(
-                "Event observer: failed to delete pending payload from database";
-                "error" => ?e
-            );
-        }
-    }
-
-    fn delete_from_db(&self, id: i64) -> Result<(), EventDispatcherError> {
-        let conn = EventDispatcherDbConnection::new_without_init(&self.db_path)
-            .expect("Failed to open database for event observer");
-        conn.delete_payload(id)?;
-        Ok(())
-    }
-
     fn send_new_attachments(&self, event_observer: &EventObserver, payload: &serde_json::Value) {
-        self.dispatch_to_observer(event_observer, payload, PATH_ATTACHMENT_PROCESSED);
+        self.dispatch_to_observer_or_log_error(event_observer, payload, PATH_ATTACHMENT_PROCESSED);
     }
 
     fn send_new_mempool_txs(&self, event_observer: &EventObserver, payload: &serde_json::Value) {
-        self.dispatch_to_observer(event_observer, payload, PATH_MEMPOOL_TX_SUBMIT);
+        self.dispatch_to_observer_or_log_error(event_observer, payload, PATH_MEMPOOL_TX_SUBMIT);
     }
 
     /// Serializes new microblocks data into a JSON payload and sends it off to the correct path
@@ -1324,7 +1269,7 @@ impl EventDispatcher {
             "burn_block_timestamp": burn_block_timestamp,
         });
 
-        self.dispatch_to_observer(event_observer, &payload, PATH_MICROBLOCK_SUBMIT);
+        self.dispatch_to_observer_or_log_error(event_observer, &payload, PATH_MICROBLOCK_SUBMIT);
     }
 
     fn send_dropped_mempool_txs(
@@ -1332,15 +1277,15 @@ impl EventDispatcher {
         event_observer: &EventObserver,
         payload: &serde_json::Value,
     ) {
-        self.dispatch_to_observer(event_observer, payload, PATH_MEMPOOL_TX_DROP);
+        self.dispatch_to_observer_or_log_error(event_observer, payload, PATH_MEMPOOL_TX_DROP);
     }
 
     fn send_mined_block(&self, event_observer: &EventObserver, payload: &serde_json::Value) {
-        self.dispatch_to_observer(event_observer, payload, PATH_MINED_BLOCK);
+        self.dispatch_to_observer_or_log_error(event_observer, payload, PATH_MINED_BLOCK);
     }
 
     fn send_mined_microblock(&self, event_observer: &EventObserver, payload: &serde_json::Value) {
-        self.dispatch_to_observer(event_observer, payload, PATH_MINED_MICROBLOCK);
+        self.dispatch_to_observer_or_log_error(event_observer, payload, PATH_MINED_MICROBLOCK);
     }
 
     fn send_mined_nakamoto_block(
@@ -1348,15 +1293,15 @@ impl EventDispatcher {
         event_observer: &EventObserver,
         payload: &serde_json::Value,
     ) {
-        self.dispatch_to_observer(event_observer, payload, PATH_MINED_NAKAMOTO_BLOCK);
+        self.dispatch_to_observer_or_log_error(event_observer, payload, PATH_MINED_NAKAMOTO_BLOCK);
     }
 
     fn send_stackerdb_chunks(&self, event_observer: &EventObserver, payload: &serde_json::Value) {
-        self.dispatch_to_observer(event_observer, payload, PATH_STACKERDB_CHUNKS);
+        self.dispatch_to_observer_or_log_error(event_observer, payload, PATH_STACKERDB_CHUNKS);
     }
 
     fn send_new_burn_block(&self, event_observer: &EventObserver, payload: &serde_json::Value) {
-        self.dispatch_to_observer(event_observer, payload, PATH_BURN_BLOCK_SUBMIT);
+        self.dispatch_to_observer_or_log_error(event_observer, payload, PATH_BURN_BLOCK_SUBMIT);
     }
 }