+++ stacks-node/src/run_loop/nakamoto.rs
@@ -1,5 +1,5 @@
 // Copyright (C) 2013-2020 Blockstack PBC, a public benefit corporation
-// Copyright (C) 2020-2023 Stacks Open Internet Foundation
+// Copyright (C) 2020-2026 Stacks Open Internet Foundation
 //
 // This program is free software: you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
@@ -77,11 +77,16 @@ pub struct RunLoop {
 
 impl RunLoop {
     /// Sets up a runloop and node, given a config.
+    ///
+    /// If no event_dispatcher is passed, a new one is created. Allowing one to be passed in
+    /// allows the nakamoto runloop to continue using the same event dispatcher as the
+    /// neon runloop at the epoch 2->3 transition.
     pub fn new(
         config: Config,
         should_keep_running: Option<Arc<AtomicBool>>,
         counters: Option<Counters>,
         monitoring_thread: Option<JoinHandle<Result<(), MonitoringError>>>,
+        event_dispatcher: Option<EventDispatcher>,
     ) -> Self {
         let channels = CoordinatorCommunication::instantiate();
         let should_keep_running =
@@ -91,11 +96,16 @@ impl RunLoop {
             config.burnchain.burn_fee_cap,
         )));
 
-        let mut event_dispatcher = EventDispatcher::new(config.get_working_dir());
-        for observer in config.events_observers.iter() {
-            event_dispatcher.register_observer(observer);
-        }
-        event_dispatcher.process_pending_payloads();
+        let event_dispatcher = event_dispatcher.unwrap_or_else(|| {
+            let mut event_dispatcher = EventDispatcher::new_with_custom_queue_size(
+                config.get_working_dir(),
+                config.node.event_dispatcher_queue_size,
+            );
+            for observer in config.events_observers.iter() {
+                event_dispatcher.register_observer(observer);
+            }
+            event_dispatcher
+        });
 
         Self {
             config,