+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to the versioning scheme outlined in the [README.md](RE
 - In the `/v3/transaction/{txid}` RPC endpoint, added `block_height` and `is_canonical` to the response.
 - New endpoint `/v3/blocks/simulate/{block_id}` allows to simulate the execution of a specific block with a brand new set of transactions
 - Improved block validation in `stacks-inspect`.
+- Allow non-blocking event dispatching. This is off by default, but can be enabled in the node configuration.
 
 ### Changed